# [c++] Optimize Varint paring for c++